Updated packages in aroma modules to org.artrev.PRODUCT_NAME
