--- aroma-core/src/main/java/io/aroma/core/Foo.java
@@ -1,5 +1,0 @@
-package io.aroma.core;
-
-public class Foo {
-
-}
