+++ aroma-core/src/main/java/org/artrev/aroma/core/Foo.java
@@ -1,0 +1,5 @@
+package org.artrev.aroma.core;
+
+public class Foo {
+
+}
