--- aroma-core/src/test/java/io/aroma/core/FooTest.java
@@ -1,10 +1,0 @@
-package io.aroma.core;
-
-import org.junit.Test;
-
-public class FooTest {
-    @Test
-    public void testSomeLibraryMethod() {
-        final Foo foo = new Foo();
-    }
-}
