+++ aroma-core/src/test/java/org/artrev/aroma/core/FooTest.java
@@ -1,0 +1,10 @@
+package org.artrev.aroma.core;
+
+import org.junit.Test;
+
+public class FooTest {
+    @Test
+    public void testSomeLibraryMethod() {
+        final Foo foo = new Foo();
+    }
+}
