+++ aroma-java-sample/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest
     xmlns:android="http://schemas.android.com/apk/res/android"
-    package="org.artrev.aroma.core"/>
+    package="org.artrev.aroma.sample"/>
