--- aroma-java-samples/build.gradle
@@ -1,23 +1,0 @@
-apply plugin: 'com.android.application'
-
-android {
-    def config = rootProject.extensions.getByName("ext")
-
-    compileSdkVersion config["androidCompileSdkVersion"]
-    buildToolsVersion config["androidBuildToolsVersion"]
-
-    defaultConfig {
-        minSdkVersion config["androidMinSdkVersion"]
-        targetSdkVersion config["androidTargetSdkVersion"]
-
-        versionCode config["versionCode"]
-        versionName config["versionName"]
-    }
-}
-
-dependencies {
-    def aromaJavaSamplesTestDependencies = rootProject.ext.aromaJavaSamplesTestDependencies
-
-    implementation project(':aroma-core')
-    testImplementation aromaJavaSamplesTestDependencies.jUnit
-}
