--- aroma-java-samples/src/test/java/io/aroma/samples/FooTest.java
@@ -1,16 +1,0 @@
-package io.aroma.samples;
-
-import org.junit.Test;
-
-import io.aroma.core.Foo;
-
-import static junit.framework.Assert.assertNotNull;
-
-public class FooTest {
-
-    @Test
-    public void shouldNotBeNull() {
-        final Foo foo = new Foo();
-        assertNotNull(foo);
-    }
-}
