+++ aroma-kotlin-ext/src/main/java/org/artrev/aroma/ext/Bar.kt
@@ -1,0 +1,3 @@
+package org.artrev.aroma.ext
+
+class Bar
