+++ aroma-kotlin-ext/src/test/java/org/artrev/aroma/ext/BarTest.kt
@@ -1,0 +1,10 @@
+package org.artrev.aroma.ext
+
+import org.junit.Test
+
+class BarTest {
+    @Test
+    fun testSomeLibraryMethod() {
+        val bar = Bar()
+    }
+}
