+++ aroma-kotlin-sample/build.gradle
@@ -1,0 +1,27 @@
+apply plugin: 'com.android.application'
+apply plugin: 'kotlin-android'
+
+android {
+    def config = rootProject.extensions.getByName("ext")
+
+    compileSdkVersion config["androidCompileSdkVersion"]
+    buildToolsVersion config["androidBuildToolsVersion"]
+
+    defaultConfig {
+        minSdkVersion config["androidMinSdkVersion"]
+        targetSdkVersion config["androidTargetSdkVersion"]
+
+        versionCode config["versionCode"]
+        versionName config["versionName"]
+    }
+}
+
+dependencies {
+    def aromaKotlinSampleDependencies = rootProject.ext.aromaKotlinSampleDependencies
+    def aromaKotlinSampleTestDependencies = rootProject.ext.aromaKotlinSampleTestDependencies
+
+    implementation project(':aroma-kotlin-ext')
+    implementation aromaKotlinSampleDependencies.kotlinStdLib
+
+    testImplementation aromaKotlinSampleTestDependencies.jUnit
+}
