+++ aroma-kotlin-sample/src/test/java/org/artrev/aroma/sample/BarTest.kt
@@ -1,0 +1,12 @@
+package org.artrev.aroma.sample
+
+import org.artrev.aroma.ext.Bar
+import org.junit.Test
+
+class BarTest {
+
+    @Test
+    fun testSomeLibraryMethod() {
+        val bar = Bar()
+    }
+}
