--- aroma-kotlin-samples/build.gradle
@@ -1,27 +1,0 @@
-apply plugin: 'com.android.application'
-apply plugin: 'kotlin-android'
-
-android {
-    def config = rootProject.extensions.getByName("ext")
-
-    compileSdkVersion config["androidCompileSdkVersion"]
-    buildToolsVersion config["androidBuildToolsVersion"]
-
-    defaultConfig {
-        minSdkVersion config["androidMinSdkVersion"]
-        targetSdkVersion config["androidTargetSdkVersion"]
-
-        versionCode config["versionCode"]
-        versionName config["versionName"]
-    }
-}
-
-dependencies {
-    def aromaKotlinSamplesDependencies = rootProject.ext.aromaKotlinSamplesDependencies
-    def aromaKotlinSamplesTestDependencies = rootProject.ext.aromaKotlinSamplesTestDependencies
-
-    implementation project(':aroma-kotlin')
-    implementation aromaKotlinSamplesDependencies.kotlinStdLib
-
-    testImplementation aromaKotlinSamplesTestDependencies.jUnit
-}
