--- aroma-kotlin-samples/src/test/java/io/aroma/samples/BarTest.kt
@@ -1,12 +1,0 @@
-package io.aroma.samples
-
-import io.aroma.kotlin.Bar
-import org.junit.Test
-
-class BarTest {
-
-    @Test
-    fun testSomeLibraryMethod() {
-        val bar = Bar()
-    }
-}
