--- aroma-kotlin/src/main/java/io/aroma/kotlin/Bar.kt
@@ -1,3 +1,0 @@
-package io.aroma.kotlin
-
-class Bar
