--- aroma-kotlin/src/test/java/io/aroma/kotlin/BarTest.kt
@@ -1,10 +1,0 @@
-package io.aroma.kotlin
-
-import org.junit.Test
-
-class BarTest {
-    @Test
-    fun testSomeLibraryMethod() {
-        val bar = Bar()
-    }
-}
